Move MultiTouchTapDetector
