+++ packages/flame/lib/input.dart
@@ -8,6 +8,6 @@ export 'src/components/input/sprite_button_component.dart';
 export 'src/events/game_mixins/multi_touch_drag_detector.dart';
 export 'src/extensions/vector2.dart';
 export 'src/game/mixins/keyboard.dart';
-export 'src/game/mixins/multi_touch_tap_detector.dart';
+export 'src/events/game_mixins/multi_touch_tap_detector.dart';
 export 'src/gestures/detectors.dart';
 export 'src/gestures/events.dart';
